test: refactor alert and snackbar utils
