+++ e2e/sms-fraud/bot-unprotected.spec.ts
@@ -2,7 +2,7 @@ import { expect, test } from '@playwright/test';
 import { TEST_ATTRIBUTES, TEST_IDS } from '../../src/client/testIDs';
 import { SMS_FRAUD_COPY } from '../../src/server/sms-fraud/smsFraudCopy';
 import { TEST_PHONE_NUMBER } from '../../src/pages/api/sms-fraud/send-verification-sms';
-import { resetScenarios } from '../resetHelper';
+import { assertAlert, assertSnackbar, resetScenarios } from '../e2eTestUtils';
 
 const TEST_ID = TEST_IDS.smsFraud;
 
@@ -19,42 +19,30 @@ test.describe('Sending verification SMS messages', () => {
     const sendButton = await page.getByTestId(TEST_ID.sendMessage);
 
     await sendButton.click();
-    const alert = await page.getByTestId(TEST_IDS.common.alert);
-    await expect(alert).toHaveAttribute(TEST_ATTRIBUTES.severity, 'success');
-    await expect(alert).toContainText(SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 2));
+    await assertAlert({ page, severity: 'success', text: SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 2) });
 
     await sendButton.click();
-    await expect(alert).toHaveAttribute(TEST_ATTRIBUTES.severity, 'error');
-    await expect(alert).toContainText(SMS_FRAUD_COPY.needToWait(1), {});
+    await assertAlert({ page, severity: 'error', text: SMS_FRAUD_COPY.needToWait(1) });
 
     await page.waitForTimeout(30000);
     await sendButton.click();
-    await expect(alert).toHaveAttribute(TEST_ATTRIBUTES.severity, 'success');
-    await expect(alert).toContainText(SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 1), {});
+    await assertAlert({ page, severity: 'success', text: SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 1) });
 
     await sendButton.click();
-    await expect(alert).toHaveAttribute(TEST_ATTRIBUTES.severity, 'error');
-    await expect(alert).toContainText(SMS_FRAUD_COPY.needToWait(2), {});
+    await assertAlert({ page, severity: 'error', text: SMS_FRAUD_COPY.needToWait(2) });
   });
 
   test('allows user to create an account with the correct code', async ({ page }) => {
     const sendButton = await page.getByTestId(TEST_ID.sendMessage);
-
     await sendButton.click();
+    await assertAlert({ page, severity: 'success', text: SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 2) });
 
-    const alert = await page.getByTestId(TEST_IDS.common.alert);
-    await expect(alert).toHaveAttribute(TEST_ATTRIBUTES.severity, 'success');
-    await expect(alert).toContainText(SMS_FRAUD_COPY.messageSent(TEST_PHONE_NUMBER, 2));
-
-    const snackBar = await page.getByTestId(TEST_IDS.common.snackBar);
-    await expect(snackBar).toHaveAttribute(TEST_ATTRIBUTES.severity, 'info');
-    await expect(snackBar).toContainText('Your verification code is');
+    await assertSnackbar({ page, severity: 'info', text: 'Your verification code is' });
     await page.getByTestId(TEST_IDS.smsFraud.copyCodeButton).click();
 
     await page.getByTestId(TEST_IDS.smsFraud.codeInput).fill(await page.evaluate('navigator.clipboard.readText()'));
     await page.getByTestId(TEST_IDS.smsFraud.sendCode).click();
-    const alert2 = await page.getByTestId(TEST_IDS.common.alert).nth(1);
-    await expect(alert2).toHaveAttribute(TEST_ATTRIBUTES.severity, 'success');
-    await expect(alert2).toContainText(SMS_FRAUD_COPY.accountCreated);
+
+    await assertAlert({ page, severity: 'success', text: SMS_FRAUD_COPY.accountCreated, index: 1 });
   });
 });
